Update key_insights_dashboard.py
Scatter plot showing the relationship between the profit and sales were created. A line chart showing the sales trend over the order date period was made and a pie chart showing sales for each customer segment was created. This concluded the charts under the Orders tab.
Under the association rules tab, a heatmap was created to measure the lift values for the antecedents and consequents.